UpdatePasswordCommand added
